+++ tcex/api/tc/v2/batch/__init__.py
@@ -1,7 +1,8 @@
 """TcEx Framework Module"""
 
 from tcex.api.tc.v2.batch.batch import Batch
+from tcex.api.tc.v2.batch.batch_cleaner import BatchCleaner
 from tcex.api.tc.v2.batch.batch_submit import BatchSubmit
 from tcex.api.tc.v2.batch.batch_writer import BatchWriter, GroupType, IndicatorType
 
-__all__ = ['Batch', 'BatchSubmit', 'BatchWriter', 'GroupType', 'IndicatorType']
+__all__ = ['Batch', 'BatchCleaner', 'BatchSubmit', 'BatchWriter', 'GroupType', 'IndicatorType']