+++ tcex/api/tc/v3/tasks/task.py
@@ -104,7 +104,7 @@ class Task(ObjectABC):
             | UserGroupModel
         ),
     ):
-        """Stage artifact on the object."""
+        """Stage assignee on the object."""
         if isinstance(data, ObjectABC):
             data = data.model  # type: ignore
         elif type.lower() == 'user' and isinstance(data, dict):